Create a python package for sss_object_detection and modify CMakeLists accordingly
