+++ .gitignore
@@ -1,3 +1,8 @@
+# Ignore data files
+*.dvs
+*.bag
+*zip
+
 # Byte-compiled / optimized / DLL files
 __pycache__/
 *.py[cod]
