+++ sss_object_detection/src/sss_object_detection/consts.py
